description of tasks

+++ src/main/java/com/ss/ita/utils/ConsoleScanner.java
@@ -93,7 +93,7 @@ public class ConsoleScanner implements Scanner {
     @Override
     public double[] readDoubleArray() {
         System.out.print("Length of the array:");
-        double[] array = new double[(int) readDouble()];
+        double[] array = new double[readInt()];
 
         int i = 0;
         while (i < array.length) {
